Add bash and batch scripts to automate module start-up
Added run.sh and run.bat to automate the start-up of all Maven sub-modules and the frontend application. This means that, instead of manually starting each component, developers can now simply run these scripts and everything will start in the correct order. This simplification makes it easier and quicker for developers to get the application running, especially those new to the project.