Update pom.xml with new plugin configurations
Updated the pom.xml file to include therapi-runtime-javadoc-scribe and lombok as dependencies under the maven-compiler-plugin. This change offers the ability to use lombok and therapi-runtime-javadoc-scribe as annotation processors. The spring-boot-maven-plugin was modified to exclude lombok. This configuration change helps in achieving more organized, reusable and maintainable code while also reducing boilerplate with the help of Lombok.