Add Codecov workflow in Github actions
A new workflow has been added to the Github Actions. This workflow is responsible for building a Java project with Maven and upload test coverage reports to Codecov. The aim is to improve the workflow execution time by caching/restoring any dependencies and to track the code coverage of test cases.