Update GH Actions workflow name to 'Codecov'
The name of the Github Actions workflow has been modified from 'Validate, Build, Test, and Package' to 'Codecov'. This change provides a more explicit understanding of the workflow's purpose, which is to build the Java project using Maven and upload the test coverage reports to Codecov. This will enhance the tracking of test case code coverage and could improve workflow execution time through dependency caching/restoration.