Make config import optional in application.properties file
The 'spring.config.import' property in the application.properties file of portfolio-service was updated. A 'classpath:application-main.properties' was set as optional. This change won't disrupt the application flow even if no application-main.properties file is present. This modification brings in more flexibility and avoids potential file not found errors.