Update Codecov upload method in workflow
Replaced Codecov GitHub action with a bash script in the CI flow. Given the increased flexibility and reliability, the new setup iterates through directories to find and upload 'jacoco.xml' reports directly to Codecov. This eliminates the need for hardcoded paths and ensures successful CI operation even in case of changes in the file structure or repository organization.