Enhance testing coverage and organization of user-service package
Expanded tests coverage to include User, Authority, and UserRegisterDTO models and introduced a test for 'contextLoads' in the UserServiceApplicationTest. This verifies the application context is being properly initialized.
Additionally, important renaming of `UserServiceUnitTest` and `UserServiceIntegrationTest` files was performed, and they were moved to a more fitting package 'service' for better organization.