Refactor portfolio service to use DTOs for responses
The portfolio service has been refactored to use Data Transfer Objects (DTOs). This enables a slightly different set of attributes to be returned than the entities used for internal operations. This makes the API more flexible as we can now easily choose what and how to present data to the user. Also, new features to get all stock assets for a user and a specific stock asset for a user have been implemented. Necessary tests have been modified to adapt to these changes.