Update Codecov settings to target main branch
This commit refines the previous setting which targeted all branches for generating coverage reports. Now, coverage reports will only be generated and uploaded to Codecov on code pushes and pull requests specifically to the main branch. It helps to focus our attention on the primary branch and reduce unnecessary resource usage.