Remove unnecessary @AllArgsConstructor from Authority and User
Removed @AllArgsConstructor annotation from Authority and User classes under auth-server package. This decision was made as the annotation was identified as unnecessary and was causing conflicts with other parts of the code. This is part of an ongoing effort to streamline the code and keep it as lean as possible. Subsequently, a test case was added for PinnacleUserPrincipal to ensure its functionality and relationships with the User and Authority entities are working as expected.